fixed signed vs unsigned vs int32 vs ssize_t warnings

+++ examples/us_example_http_server.c
@@ -125,7 +125,7 @@ int us_webapp_json_test_dispatch(
     (void)request_content;
     sprintf( json, "{ \"time\" : \%ld, \"access_count\" = %d }", (long)time(0), (int)cnt++ );
     r&=us_buffer_append_string( response_content, json );
-    r&=us_http_response_header_init_ok( response_header, 200, "application/json", us_buffer_readable_count( response_content ),true);
+    r&=us_http_response_header_init_ok( response_header, 200, "application/json", (ssize_t)us_buffer_readable_count( response_content ),true);
     r&=us_http_response_header_set_no_cache( response_header );
     if ( r )
         return response_header->m_code;
